Implement generic operator type (#265)
* Switch from gem to mul

* Implement generalized Operator type

* Allow arbitrary array types as data

* Add some tests for sparse density matrices

* Loosen type constraints in dmaster_nh to work with lazy adjoints

* Add tests for irreducibility checks

* Better typing for alpha and beta in dmaster

* Infer types for alpha and beta in derivative functions

* Use eltype for types of alpha and beta in derivatives

* Better typing in stochastics

* Loosen type constraints on tspan

* Update version and dependencies

+++ src/stochastic_definitions.jl
@@ -46,18 +46,18 @@ function homodyne_carmichael(H0::AbstractOperator, C::Vector{T}, theta::Vector{R
     X = C .* exp.(-1.0im .* theta) + dagger.(C) .* exp.(1.0im .* theta)
     CdagC = -0.5im .* dagger.(C) .* C
 
-    fstoch(t::Float64, psi::StateVector) = Hs
+    fstoch(t, psi::StateVector) = Hs
     if normalize_expect
         function H_nl_n(psi::StateVector)
             psi_n = normalize(psi)
             sum(expect(X[i], psi_n)*Hs[i] + CdagC[i] for i=1:n)
         end
-        fdeterm_n(t::Float64, psi::StateVector) = H0 + H_nl_n(psi)
+        fdeterm_n(t, psi::StateVector) = H0 + H_nl_n(psi)
         return fdeterm_n, fstoch
     else
         H_nl_un(psi::StateVector) =
             sum(expect(X[i], psi)*Hs[i] + CdagC[i] for i=1:n)
-        fdeterm_un(t::Float64, psi::StateVector) = H0 + H_nl_un(psi)
+        fdeterm_un(t, psi::StateVector) = H0 + H_nl_un(psi)
         return fdeterm_un, fstoch
     end
 end
